refactor: mejorar nombres de carpetas y archivos respetando la convencion
