+++ src/main/java/org/example/c2/Ejercicio.java
@@ -1,8 +1,8 @@
-package org.example.clase_2;
+package org.example.c2;
 
 import java.util.Scanner;
 
-public class Ejercicio_clase_2 {
+public class Ejercicio {
     public static void main(String[] args) {
         // ----------------------------
         // 1. Variables y Operadores
