+++ src/main/java/org/example/c3/diagrama.png
